+++ audiobook/src/main/java/de/ph1b/audiobook/features/book_overview/BookShelfFragment.kt
@@ -100,7 +100,7 @@ class BookShelfFragment : RxBaseFragment<BookShelfFragment, BookShelfBasePresent
 
         // init fab
         fab.setIconDrawable(playPauseDrawable)
-        fab.setOnClickListener { presenter()!!.playPauseRequested() }
+        fab.setOnClickListener { presenter().playPauseRequested() }
 
         // init ActionBar
         actionBar().apply {