+++ audiobook/src/main/java/de/ph1b/audiobook/mvp/RxBaseActivity.kt
@@ -35,7 +35,7 @@ abstract class RxBaseActivity<V, P> : BaseActivity() where P : Presenter<V> {
 
     abstract fun provideView(): V
 
-    fun presenter() = presenterDelegate.presenter
+    fun presenter() = presenterDelegate.presenter()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)